Create a mock file
