+++ modules/fee-market/src/lib.rs
@@ -24,6 +24,8 @@
 #[cfg(feature = "runtime-benchmarks")]
 mod benchmarking;
 #[cfg(test)]
+mod mock;
+#[cfg(test)]
 mod tests;
 
 pub mod weight;
